add first e2e test and infra around it

+++ internal/pkg/testconn/postgres/conn.go
@@ -7,6 +7,7 @@ import (
 	"testing"
 
 	"github.com/viktorkomarov/datagen/internal/model"
+	"github.com/viktorkomarov/datagen/internal/pkg/testconn/options"
 	"github.com/viktorkomarov/datagen/internal/pkg/xrand"
 
 	"github.com/jackc/pgx/v5"
@@ -14,7 +15,8 @@ import (
 )
 
 type Conn struct {
-	conn *pgx.Conn
+	saveSchema bool
+	conn       *pgx.Conn
 }
 
 func New(t *testing.T, connStr string) (*Conn, error) {
@@ -47,15 +49,19 @@ func New(t *testing.T, connStr string) (*Conn, error) {
 		return nil, fmt.Errorf("%w: connect config", err)
 	}
 
+	c := &Conn{conn: tempConn, saveSchema: false}
+
 	t.Cleanup(func() {
 		ctx := context.Background()
 		if err = tempConn.Close(ctx); err != nil {
 			t.Errorf("failed to close temp conn: %v", err)
 		}
 
-		_, err := conn.Exec(ctx, fmt.Sprintf("DROP DATABASE %s WITH (force)", dbname))
-		if err != nil {
-			t.Errorf("failed to drop database %s: %v", dbname, err)
+		if !c.saveSchema {
+			_, err := conn.Exec(ctx, fmt.Sprintf("DROP DATABASE %s WITH (force)", dbname))
+			if err != nil {
+				t.Errorf("failed to drop database %s: %v", dbname, err)
+			}
 		}
 
 		if err = conn.Close(ctx); err != nil {
@@ -63,7 +69,7 @@ func New(t *testing.T, connStr string) (*Conn, error) {
 		}
 	})
 
-	return &Conn{conn: tempConn}, nil
+	return c, nil
 }
 
 func createTempDB(ctx context.Context, conn *pgx.Conn) (string, error) {
@@ -80,50 +86,26 @@ func (c *Conn) Raw() *pgx.Conn {
 	return c.conn
 }
 
-type partPolicy struct {
-	Method string
-	Cnt    int
-	Field  string
-}
-
-type CreateTableOptions struct {
-	pks        []string
-	partPolicy partPolicy
-}
-
-type CreateTableOption func(c *CreateTableOptions)
-
-func WithPKs(pks []string) CreateTableOption {
-	return func(c *CreateTableOptions) {
-		c.pks = pks
-	}
-}
-
-func WithHashPartitions(parts int, field string) CreateTableOption {
-	return func(c *CreateTableOptions) {
-		c.partPolicy = partPolicy{
-			Method: "hash",
-			Cnt:    parts,
-			Field:  field,
-		}
-	}
-}
-
-func (c *Conn) CreateTable(ctx context.Context, table model.Table, opts ...CreateTableOption) error {
+func (c *Conn) CreateTable(ctx context.Context, table model.Table, opts ...options.CreateTableOption) error {
 	if err := c.ensureUnexistence(ctx, table.Name); err != nil {
 		return fmt.Errorf("%w: ensure unexistence %s", err, table.Name)
 	}
 
-	params := CreateTableOptions{
-		pks: make([]string, 0),
-		partPolicy: partPolicy{
+	params := options.CreateTableOptions{
+		PKs: make([]string, 0),
+		PartPolicy: options.PartPolicy{
 			Method: "",
 			Cnt:    0,
 			Field:  "",
 		},
+		Preserve: false,
 	}
 	for _, opt := range opts {
 		opt(&params)
+	}
+
+	if params.Preserve {
+		c.saveSchema = true
 	}
 
 	query := fmt.Sprintf("create table %s (", table.Name.String())
@@ -131,22 +113,22 @@ func (c *Conn) CreateTable(ctx context.Context, table model.Table, opts ...Creat
 		return fmt.Sprintf("%s %s", c.Name, c.Type)
 	}), ",")
 
-	if len(params.pks) != 0 {
-		query += fmt.Sprintf(",primary key (%s)", strings.Join(params.pks, ","))
+	if len(params.PKs) != 0 {
+		query += fmt.Sprintf(",primary key (%s)", strings.Join(params.PKs, ","))
 	}
 
 	query += ")"
 
-	if params.partPolicy.Method != "" {
-		query += fmt.Sprintf("partition by hash(%s)", params.partPolicy.Field)
+	if params.PartPolicy.Method != "" {
+		query += fmt.Sprintf("partition by hash(%s)", params.PartPolicy.Field)
 		if _, err := c.conn.Exec(ctx, query); err != nil {
 			return fmt.Errorf("%w: create table", err)
 		}
 
-		for i := range params.partPolicy.Cnt {
+		for i := range params.PartPolicy.Cnt {
 			part := fmt.Sprintf(
 				"create table %s_part_%d partition of %s for values with (modulus %d, remainder %d)",
-				table.Name.String(), i, table.Name.String(), params.partPolicy.Cnt, i,
+				table.Name.String(), i, table.Name.String(), params.PartPolicy.Cnt, i,
 			)
 
 			if _, err := c.conn.Exec(ctx, part); err != nil {
@@ -159,6 +141,39 @@ func (c *Conn) CreateTable(ctx context.Context, table model.Table, opts ...Creat
 
 	if _, err := c.conn.Exec(ctx, query); err != nil {
 		return fmt.Errorf("%w: create table", err)
+	}
+
+	return nil
+}
+
+func (c *Conn) OnEachRow(ctx context.Context, table model.Table, fn func(row []any)) error {
+	columns := lo.Map(table.Columns, func(c model.Column, _ int) string {
+		return string(c.Name)
+	})
+	query := "SELECT " + strings.Join(columns, ", ") + " FROM " + table.Name.String()
+
+	rows, err := c.conn.Query(ctx, query)
+	if err != nil {
+		return fmt.Errorf("%w: on each row", err)
+	}
+	defer rows.Close()
+
+	buf := make([]any, len(table.Columns))
+	ptrBuf := make([]any, len(buf))
+	for i := range ptrBuf {
+		ptrBuf[i] = &buf[i]
+	}
+
+	for rows.Next() {
+		if err := rows.Scan(ptrBuf...); err != nil {
+			return fmt.Errorf("%w: on each row", err)
+		}
+
+		fn(buf)
+	}
+
+	if err = rows.Err(); err != nil {
+		return fmt.Errorf("%w: on each row", err)
 	}
 
 	return nil
